deploy/requirements.txt
fastapi==0.112.1
uvicorn==0.30.6
pandas==2.2.2
google-cloud-bigquery==3.25.0
db-dtypes==1.3.0